# openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Atlas API
  version: '1'
  description: |
    Atlas is the Northeastern University Library digital repository API.

    Two consumption surfaces of this document:

    - Humans: Scalar UI at `/docs`
    - Machines (clients, agents, codegen): the canonical OpenAPI
      document is served at `/api-docs/openapi.yaml`.

    ### Authentication

    Most write endpoints require a Cerberus system bearer token plus a
    `User` header carrying the acting NUID. Read endpoints generally
    allow unauthenticated guest access. Endpoints that require auth
    declare it explicitly below.
servers:
- url: "{scheme}://{host}"
  variables:
    scheme:
      default: https
      enum:
      - https
      - http
    host:
      default: atlas.library.northeastern.edu
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Cerberus system token or a devise-jwt user token. Both are accepted.
    NuidHeader:
      type: apiKey
      in: header
      name: User
      description: NUID identity asserted by Cerberus, in the form `NUID <nuid>`.
        Required when calling with the system token to act as a user.
    OnBehalfOfHeader:
      type: apiKey
      in: header
      name: On-Behalf-Of
      description: 'Acting-as attribution target, in the form `NUID <nuid>`. Sent
        on writes during an acting-as session: the `User` header (the operator) authorizes
        the request and must be an admin; this header names the user the write is
        attributed to. On a create, the resulting resource reads as a pure deposit
        by the target — `depositor` = target, `proxy_uploader` left null — with the
        operator recorded only in the AuditEvent. A non-admin operator presenting
        this header is rejected (403).'
  schemas:
    Work:
      type: object
      properties:
        work:
          type: object
          properties:
            id:
              type: string
              description: NOID
            valkyrie_id:
              type: string
              description: Valkyrie internal id
            ancestors:
              type: array
              items:
                type: array
                items:
                  type: string
                minItems: 2
                maxItems: 2
                description: "[noid, type-name] pair"
            ancestor_chain:
              type: array
              description: Ancestor chain, root-first, with titles — array of {noid,
                klass, title} objects
              items:
                type: object
                properties:
                  noid:
                    type: string
                    description: Ancestor NOID
                  klass:
                    type: string
                    description: Ancestor resource class name
                  title:
                    type: string
                    description: Ancestor plain-text title (may be empty)
                required:
                - noid
                - klass
                - title
            thumbnail:
              type: string
              nullable: true
              description: IIIF URL of the :thumbnail_image Delegate (~85px), or null
            thumbnail_2x:
              type: string
              nullable: true
              description: IIIF URL of the :thumbnail_image_2x Delegate (~170px retina),
                or null
            preview:
              type: string
              nullable: true
              description: IIIF URL of the :preview_image Delegate (~500px hero),
                or null
            title:
              type: string
              nullable: true
            description:
              type: string
              nullable: true
            permanent_url:
              type: string
              nullable: true
            tombstoned:
              type: boolean
              description: Withdrawn-from-discovery flag
            tombstoned_at:
              type: string
              nullable: true
              description: ISO-8601 timestamp set when tombstoned
            tombstoned_by:
              type: string
              nullable: true
              description: NUID of the user who tombstoned the resource
            depositor:
              type: string
              nullable: true
              description: NUID of the intellectual owner (the named depositor; may
                differ from the hands-on-keyboard actor)
            proxy_uploader:
              type: string
              nullable: true
              description: NUID of the hands-on-keyboard actor for the most-recent
                create. Equals depositor for self-deposit; differs in the librarian-on-behalf
                (proxy-deposit) case. Left null under acting-as impersonation (On-Behalf-Of
                header present) — the operator is recorded only in the AuditEvent,
                not on the resource.
            in_progress:
              type: boolean
              description: Cerberus-driven workflow flag; true until the bulk-deposit
                job marks the Work complete.
            incomplete:
              type: boolean
              description: Pipeline-failure flag; true when a work-scoped enrichment
                job gave up after its retries. Flags only — the Work stays readable.
            incomplete_reason:
              type: string
              nullable: true
              description: Machine token naming the cause (e.g. pdf_rendition_gave_up).
                Opaque to Atlas and unvalidated; the vocabulary belongs to the caller.
                Null unless the Work is flagged.
            derivative_permissions:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
              description: 'Per-asset derivative read policy: sparse map of tier =>
                [read groups]. Image ladder small/medium/large/service/master plus
                independent media audio/video/pdf. Empty when unset (tiers inherit
                the Work visibility).'
          required:
          - id
          - valkyrie_id
          - ancestors
          - ancestor_chain
          - thumbnail
          - thumbnail_2x
          - preview
          - title
          - description
          - permanent_url
          - tombstoned
          - tombstoned_at
          - tombstoned_by
          - depositor
          - proxy_uploader
          - in_progress
          - incomplete
          - incomplete_reason
          - derivative_permissions
      required:
      - work
...
paths:
...
  "/works/{id}":
    parameters:
    - name: id
      in: path
      description: NOID of the Work
      required: true
      schema:
        type: string
    get:
      summary: Retrieve a work
      tags:
      - Works
      responses:
        '200':
          description: ancestor_chain carries each ancestor noid, klass and title
            (root-first)
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Work"
        '410':
          description: work tombstoned
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Work"
    patch:
      summary: Update a work
      tags:
      - Works
      description: |
        Update a Work's descriptive metadata by supplying a `binary` MODS XML
        upload — the caller assembles the full document (descriptive merge logic
        lives in the client, e.g. Cerberus, not Atlas). `metadata[permissions]`
        adjusts the ACL. Any `metadata[title]` / `metadata[description]` keys are
        ignored.

        Two rules bound an ACL write. A resource may be no more visible than its
        container, so a read audience wider than the parent's is refused with
        `422 visibility_exceeds_parent` (widen the parent instead). And a group
        grant may only be removed by a member of that group — admin and the
        devolved-admin tier excepted; a grant the caller cannot remove is
        preserved rather than rejected, so the stored ACL may retain groups the
        request omitted. Read it back with `GET /resources/{id}/permissions`.

        Programmatic Delegate writes (thumbnail-family URIs, sized image
        derivatives) no longer ride this endpoint — see the dedicated
        `PATCH /works/{id}/thumbnails` and `PATCH /works/{id}/image_derivatives`
        routes.
      parameters: []
      requestBody:
        required: false
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                binary:
                  type: string
                  format: binary
                  description: MODS XML to apply to the Work
      responses:
        '200':
          description: work updated
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Work"
        '409':
          description: optimistic-lock conflict on the update (surfaced immediately,
            not retried)
    delete:
      summary: Destroy a work
      tags:
      - Works
      description: |
        Permanently removes the Work. This is a purge, not a withdrawal: it
        deletes the Work's metadata, cascades into its FileSets and their
        Blobs, and removes the OCFL objects that hold the preserved bytes —
        every retained revision, not only the current one. Nothing survives
        but the audit row, which records the NOIDs it removed.

        Use `POST /works/{id}/tombstone` for the user-visible withdrawal
        path. That one keeps everything and can be reversed.

        Admin only.
      responses:
        '204':
          description: work destroyed
...